The most significant changes include the addition of a new section in the README.md file, improvements in code readability and maintainability in the GLContext.java file, and updates in the RenderStreamINTEL.java, RenderStreamPBO.java, StreamBuffered.java, and StreamBufferedPBO files. The changes in the StreamUtil, TextureStreamINTEL, TextureStreamPBO, TextureStreamPBODefault, and TextureStreamPBORange classes are also noteworthy.
1. The README.md file was updated with a new section titled "构建" and instructions on how to rewrite the `MANIFEST.MF` file of `org.lwjgl.lwjgl：lwjgl` to set `Sealed: false`.
2. The GLContext.java file was updated for better code readability and maintainability. This includes adding return statements for methods, breaking down complex lines of code into multiple lines, and replacing some method calls with variable assignments.
3. The RenderStream.java and GLContext.java files were updated to move the closing braces of the classes to new lines for consistency.
4. The RenderStreamINTEL.java file was updated with new variables and some method calls were replaced with variable assignments. Some error messages were also updated to use string interpolation.
5. The RenderStreamPBO.java file was updated similarly to the RenderStreamINTEL.java file.
6. The RenderStreamPBOAMD.java, RenderStreamPBOCopy.java, and RenderStreamPBODefault.java files were updated to use long integers for the size parameter in the `glBufferData` and `glMapBuffer` methods.
7. The StreamBuffered.java file was updated with new variables and a new method `waitForProcessingToComplete`.
8. The `StreamBufferedPBO` class was updated to include a loop that generates and binds buffers for each element in the `pbos` array.
9. The `StreamHandler` interface was not changed, only the formatting of the closing bracket was altered.
10. The `StreamUtil` class was updated in several places, including the `createRenderTexture` and `createRenderBuffer` methods, the `getStride` method, the `checkCapabilities` method, the `getRenderStreamImplementation` method, the `getRenderStreamImplementations` method, and the `checkSamples` method.
11. The `TextureStream` interface was not changed, only the formatting of the closing bracket was altered.
12. The `TextureStreamINTEL` class was updated in several places, including the `strideBuffer` and `layoutBuffer`, the `genLayoutLinearTexture` method, the `resize` method, the `snapshot` method, the `tick` method, and the `destroyObjects` method.
13. The `TextureStreamPBO` class was updated in several places, including the `resize` method, the `snapshot` method, the `tick` method, the `upload` method, and the `destroyObjects` method.
14. The `TextureStreamPBODefault` class was updated to cast the height and stride to a long when calling `glBufferData` and `glMapBuffer` in the `pinBuffer` method.
15. The `TextureStreamPBORange` class was updated to include an array of `GLSync` objects called `fences`. The `pinBuffer` method was also updated.

+++ src/main/java/org/lwjgl/util/stream/RenderStreamPBODefault.java
@@ -18,33 +18,35 @@ final class RenderStreamPBODefault extends RenderStreamPBO {
         public RenderStream create(final StreamHandler handler, final int samples, final int transfersToBuffer) {
             final ContextCapabilities caps = GLContext.getCapabilities();
             return new RenderStreamPBODefault(
-                    handler, samples, transfersToBuffer,
-                    StreamUtil.isNVIDIA(caps) ? ReadbackType.GET_TEX_IMAGE : ReadbackType.READ_PIXELS
+                handler, samples, transfersToBuffer,
+                StreamUtil.isNVIDIA(caps) ? ReadbackType.GET_TEX_IMAGE : ReadbackType.READ_PIXELS
             );
         }
     };
+
     private final boolean USE_COPY_BUFFER_SUB_DATA;
 
     RenderStreamPBODefault(final StreamHandler handler, final int samples, final int transfersToBuffer, final ReadbackType readbackType) {
         super(handler, samples, transfersToBuffer, readbackType);
         final ContextCapabilities caps = GLContext.getCapabilities();
+
         USE_COPY_BUFFER_SUB_DATA = (caps.OpenGL31 || caps.GL_ARB_copy_buffer) &&
-                !StreamUtil.isAMD(caps);
+            !StreamUtil.isAMD(caps);
     }
 
     protected void pinBuffer(final int index) {
         glBindBuffer(GL_PIXEL_PACK_BUFFER, pbos[index]);
-        pinnedBuffers[index] = glMapBuffer(GL_PIXEL_PACK_BUFFER, GL_READ_ONLY, height * stride, pinnedBuffers[index]);
+        pinnedBuffers[index] = glMapBuffer(GL_PIXEL_PACK_BUFFER, GL_READ_ONLY, (long) height * stride, pinnedBuffers[index]);
         glBindBuffer(GL_PIXEL_PACK_BUFFER, 0);
     }
 
     protected void copyFrames(final int src, final int trg) {
         if (USE_COPY_BUFFER_SUB_DATA) {
             glBindBuffer(GL_COPY_WRITE_BUFFER, pbos[trg]);
-            glCopyBufferSubData(GL_PIXEL_PACK_BUFFER, GL_COPY_WRITE_BUFFER, 0, 0, height * stride);
+            glCopyBufferSubData(GL_PIXEL_PACK_BUFFER, GL_COPY_WRITE_BUFFER, 0, 0, (long) height * stride);
             glBindBuffer(GL_COPY_WRITE_BUFFER, 0);
         } else {
-            pinnedBuffers[src] = glMapBuffer(GL_PIXEL_PACK_BUFFER, GL_READ_ONLY, height * stride, pinnedBuffers[src]);
+            pinnedBuffers[src] = glMapBuffer(GL_PIXEL_PACK_BUFFER, GL_READ_ONLY, (long) height * stride, pinnedBuffers[src]);
             glBindBuffer(GL_PIXEL_PACK_BUFFER, pbos[trg]);
             glBufferSubData(GL_PIXEL_PACK_BUFFER, 0, pinnedBuffers[src]);
             glBindBuffer(GL_PIXEL_PACK_BUFFER, pbos[src]);
